requirements.txt
asn1crypto==1.3.0
billiard==3.6.1.0
# celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0.2
django-configurations==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.11.0
future==0.18.2
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
more-itertools==8.2.0
psycopg2
psycopg2-binary==2.8.4
django-decouple
django-cors-headers==3.2.1
drf-yasg==1.17.0
restframework-simplejwt==4.4.0
celery==4.3.0
django-filter==2.2.0
django-debug-toolbar==2.1
Pillow==7.0.0
redis==3.3.11